WIP: separate main function to another file
